feat: add `User`, `Company` and update `Quote`
Current state of application a little bit complex now, which
`Quote` is belongs to `Company`, `User` is belongs to `Company`,
and `Company` has many `User` and `Quote`.

With this association, there is a security issues related to the
`Turbo Streams`. Therefore, the broadcast from `Quote` and
the turbo stream in `quotes#index` are update to make it uniq.